requirements.txt
instagrapi
google-genai
requests
pillow
python-dotenv
